[DOES NOT COMPILE] swithcing PGD and tracer model to vec3 and float

+++ src/plume/Particle.h
@@ -201,8 +201,8 @@ public:
   // double vTot{};
   // double wTot{};
 
-  double CoEps{};
-  double nuT{};
+  float CoEps{};
+  float nuT{};
 
   // The velocity fluctuation for a particle from the last iteration
   double uFluct_old{};// u component
